requirements.txt
google-ads>=28.2.0
pandas>=2.0.0
google-auth-oauthlib>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

